backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
requests==2.31.0
pydantic==2.5.2
pydantic-settings==2.1.0
funasr>=1.0.0
numpy>=1.24.0
ffmpeg-python>=0.2.0
